tools/pyproject.toml
[project]
name = "astro-blog-tools"
version = "0.1.0"
description = "Automation tools for Servando's Astro blog"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "instructor>=1.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "dub>=0.23.1",
    "python-dotenv>=1.0.0",
    "elevenlabs>=1.0.0",
    "markdownify>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "COM812", "ISC001"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true